Labyrinth first push
Cameron:
Pushing my labyrinth project.  It has basic camera/movement with some
public variables to mess with and a stamina system.  The labyrinth Matt
made is in as well, so you might need blender in order to open the
project.  Also included is my first AI script.  It's got
find/investigate/sighted modes with a rudimentary "sight" system using
raycasts.  I think this is a good base to build off of.