Start menu fix and lighting bake
Reworked the menu script to play the level--Application.loadlevel was
being called on start which I guess freaked Unity out.  I also baked the
lighting on the maze so if there are any changes to it we'll need to
remember to rebake it to apply the lighting changes.  More folder
organization too.