Fixed Labyrinth
There was a small glitch on one of the walls of the labyrinth where you
would be able to essentially jump onto the top of the labyrinth walls
because the wall in question was "inside out". That no longer happens.
It wasn't like anyone actually knew this was a problem in the first
place, so, whatever. It's fixed now though.