Small Graphical Changes
Added lever mesh and texture. Placed it behind the starting position in
the corner. Added an orange light to the exit, called ExitSpotlight.
Textured the minotaur, it looks terrible, I know, you don't have to tell
me. Mainly its just so there isn't an object that is just stark white,
and also I wanted to mess with texturing in blender.