end credits and slight renaming
Now when the player reaches the end of the maze, a simple scrolling
credits occurs. After our names scroll through, it says thanks for
playing and it blinks "click to continue". Currently, this leads back to
the start screen. Additionally, renamed the "cylinder", our initial item
to pick up to Lever Piece.