So here's a backstory
Judge my limerick. Currently it's set to show for 5 seconds before
loading the main scene. I do not have it connected to the main menu
script as of right now. I was having issues doing the fade in with the
main scene, so feel free to comment to me or whatever.