model for sound triggers
Walk straight and to the left. The trigger for audio is by the light,
set up with an invisible cube. RIP headphone users.

Hi everyone, my name is Jason, and sometimes I type collision instead of
trigger, causing me to be unable to figure out why my sound isn't
working and Google solutions about audiosources for three hours, when in
actuality, my code was correct with exception of the method name.