Vines and Mesh Fixes
When I push this I fully expect everything to break and to have an error
with Github. If you guys see this, then please warn me when you are
going to push to Github. I am always making changes within the game and
will need to push before you do so that the main scene doesn't reset all
of my hard work.