idk
it is midnight and we just want to go to sleep. Unity, we beg you.

There was an issue where the animation would play immediately for some
reason. This was attributed to nothing as it turns out, but what was an
actual problem was 2 vital objects were both named "ConnectionPiece".
One of those objects is now named "Disc" while the other retains its
name.