High scores updates and final credits
The timer was removed and can no longer be seen as the player moves
through the labyrinth. Due to this, I added a new scene called
"GameScore," which displays the number of deaths and total time it took
for the player to win.

Additionally, if the player makes it to the high scores, the text of
their score is now red, instead of white.

For the end credits, I included the name of the guy that made the pieces
of eight font.

Happy Thanksgiving, everybody!